Reorganize the code structure to make the current design of the cognitive module clearer.
